refactor: new components for nav bar
